requirements.txt
numpy==2.3.3
pillow==11.3.0
customtkinter==5.2.2
opencv-python==4.12.0.88
ultralytics==8.3.197
picamera2==0.3.30
simplejpeg==1.8.2
shapely==2.1.1
matplotlib==3.10.6
PySide6==6.8.0.2
qt-material==2.1.0
